+++ src/heap/mod.rs
@@ -57,6 +57,8 @@
 //!     metadata => pub Heap;
 //!     // Use this heap as the global allocator.
 //!     global => true;
+//!     // Uncomment the following line to enable heap tracing feature:
+//!     // trace_port => 31;
 //! }
 //!
 //! // Create a static instance of the heap type and declare it as the global